// codegen/positions/Positions.types.ts
export type Decimal = string;
export type Uint128 = string;
export interface NativeToken {
    denom: string;
};
export type Addr = string;
export interface BasketPositionsResponse {
  positions: PositionResponse[];
  user: string;
}
export interface PositionResponse {
  avg_borrow_LTV: Decimal;
  avg_max_LTV: Decimal;
  cAsset_ratios: Decimal[];
  collateral_assets: CAsset[];
  credit_amount: Uint128;
  position_id: Uint128;
}
export interface CAsset {
  asset: Asset;
  max_LTV: Decimal;
  max_borrow_LTV: Decimal;
  pool_info?: PoolInfo | null;
  rate_index: Decimal;
}
export interface Asset {
  amount: Uint128;
  info: NativeToken;
}
export interface PoolInfo {
  asset_infos: LPNativeToken[];
  pool_id: number;
}
export interface LPNativeToken {
  decimals: number;
  info: NativeToken;
  ratio: Decimal;
}
export interface Basket {
  base_interest_rate: Decimal;
  basket_id: Uint128;
  collateral_supply_caps: SupplyCap[];
  collateral_types: CAsset[];
  cpc_margin_of_error: Decimal;
  credit_asset: Asset;
  credit_last_accrued: number;
  credit_price: PriceResponse;
  current_position_id: Uint128;
  frozen: boolean;
  lastest_collateral_rates: Rate[];
  liq_queue?: Addr | null;
  multi_asset_supply_caps: MultiAssetSupplyCap[];
  negative_rates: boolean;
  oracle_set: boolean;
  pending_revenue: Uint128;
  rates_last_accrued: number;
  rev_to_stakers: boolean;
}
export interface SupplyCap {
  asset_info: NativeToken;
  current_supply: Uint128;
  debt_total: Uint128;
  lp: boolean;
  stability_pool_ratio_for_debt_cap?: Decimal | null;
  supply_cap_ratio: Decimal;
}
export interface PriceResponse {
  decimals: number;
  price: Decimal;
  prices: PriceInfo[];
}
export interface PriceInfo {
  price: Decimal;
  source: string;
}
export interface Rate {
  last_time_updated: number;
  rate: Decimal;
}
export interface MultiAssetSupplyCap {
  assets: NativeToken[];
  supply_cap_ratio: Decimal;
}
export interface CollateralInterestResponse {
  rates: Decimal[];
}
// ...
export interface InterestResponse {
  credit_interest: Decimal;
  negative_rate: boolean;
}